requirements.txt
solace-pubsubplus
python-dotenv
pymongo
defusedxml
certifi
fastapi
uvicorn
